Refactor : handshakeinterceptor 패키지를 util 패키지로 변경
handshakeinterceptor의 용도보다 util용도가 강해져 패키지 명을 변경합니다.
