+++ src/main/java/nbbang/com/nbbang/global/socket/StompWebSocketMessageBrokerConfig.java
@@ -1,7 +1,7 @@
 package nbbang.com.nbbang.global.socket;
 
 import lombok.RequiredArgsConstructor;
-import nbbang.com.nbbang.global.socket.handshakeinterceptor.LoginHandShakeInterceptor;
+import nbbang.com.nbbang.global.socket.util.CustomHandShakeInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
@@ -15,7 +15,7 @@ import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerCo
 public class StompWebSocketMessageBrokerConfig implements WebSocketMessageBrokerConfigurer {
 
     private final StompChannelInterceptor stompChannelInterceptor;
-    private final LoginHandShakeInterceptor loginInterceptor;
+    private final CustomHandShakeInterceptor loginInterceptor;
 
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
